Add ruute for new position of BookForm.js and add style fiel
