--- src/components/BookForm.js
@@ -1,53 +1,0 @@
-import React, { useState } from 'react';
-import { useDispatch } from 'react-redux';
-import { nanoid } from '@reduxjs/toolkit';
-import { createBook } from '../redux/books/booksSlice';
-
-function BookForm() {
-  const [title, setTitle] = useState('');
-  const [author, setAuthor] = useState('');
-  const [category, setCategory] = useState('');
-  const dispatch = useDispatch();
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    if (title && author && category) {
-      const newBook = {
-        item_id: nanoid(),
-        title,
-        author,
-        category,
-      };
-      dispatch(createBook(newBook));
-      setTitle('');
-      setAuthor('');
-      setCategory('');
-    }
-  };
-
-  return (
-    <form onSubmit={handleSubmit}>
-      <input
-        type="text"
-        placeholder="Title"
-        value={title}
-        onChange={(e) => setTitle(e.target.value)}
-      />
-      <input
-        type="text"
-        placeholder="Author"
-        value={author}
-        onChange={(e) => setAuthor(e.target.value)}
-      />
-      <input
-        type="text"
-        placeholder="Category"
-        value={category}
-        onChange={(e) => setCategory(e.target.value)}
-      />
-      <button type="submit">Add Book</button>
-    </form>
-  );
-}
-
-export default BookForm;
